:sparkles: Add dependencies to phases
We need to be able to specify that tasks have dependencies. In PLA-land,
this means that a task that has one or more "dep" statements indicates that
those tasks defined in the "dep" statement are DEPENDENT on the task
containing the "dep" statements.

This seems backwards, but it is how the file format is defined, so C'est la
vie.