:sparkles: Make RecipeSpecs use a default start date if not provided
RecipeSpec objects previously always required a start_date, as a string,
when deserializing. This is not the behavior we want, though. Instead, what
we want is for the start string for RecipeSpecs to be optional, and if not
provided, to use the start date in the timeline of the overall
ProductionSchedule.