:sparkles: Add start dates to recipe instances
Instances of the data structure Recipe need to have start dates associated
with them. Further, this start date should be computed for each phase
instance, as it's being converted to PLA format. This allows us to output
start date(s)/times for phases in order to produce the basic Gantt chart.