:sparkles: Add command line arguments for input and output to main.rs
Eventually, we'll need to take input from more than just standard input,
and allow for output to more than just standard output. In order to
accomplish this task, we need to make sure that the command line arguments
for the binary version of chronogrog take options that allow the user to
specify these arguments.

Refs #2.