:sparkles: Defer PhaseInstances until resources are free.
We need to be able to defer the start of a PhaseInstance until necessary
Resource objects are free. This commit restructures the Resource object so
that it contains a vector of periods which it is allocated. In addition,
the interface to ResourceTracker has slightly changed, now querying
individual Resource objects whether or not they are free for a given
period.

Refs #3.